+++ src/proton/types/update/delete_messages.cr
@@ -8,10 +8,10 @@ module Proton::Types
     include JSON::Serializable
 
     # Chat identifier
-    property chat_id : ::Int64
+    property chat_id : ::Int32
 
     # Identifiers of the deleted messages
-    property message_ids : ::Array(::Int64)
+    property message_ids : ::Array(::Int32)
 
     # True, if the messages are permanently deleted by a user (as opposed to just becoming inaccessible)
     property is_permanent : ::Bool
@@ -19,7 +19,7 @@ module Proton::Types
     # True, if the messages are deleted only from the cache and can possibly be retrieved again in the future
     property from_cache : ::Bool
 
-    def initialize(@chat_id : ::Int64, @message_ids : ::Array(::Int64), @is_permanent : ::Bool, @from_cache : ::Bool)
+    def initialize(@chat_id : ::Int32, @message_ids : ::Array(::Int32), @is_permanent : ::Bool, @from_cache : ::Bool)
     end
   end
 end
